have beginnings of mach-o support.
So far, can parse header (not doing anything with cpu subtype)
Next steps are to implement load_commands

+++ capstone/src/instruction.rs
@@ -47,8 +47,8 @@ pub struct Instruction<'a> {
     pub mnemonic: &'a str,
     pub op_str: &'a str,
     pub groups: Option<&'a [u8]>,
-    pub regs_read: Option<&'a [u8]>,
-    pub regs_write: Option<&'a [u8]>,
+    pub regs_read: Option<&'a [u16]>,
+    pub regs_write: Option<&'a [u16]>,
     pub detail: Option<&'a capstone_sys::bindings::cs_detail__bindgen_ty_1>,
 }
 impl<'a> Instruction<'a> {
@@ -63,8 +63,8 @@ impl<'a> Instruction<'a> {
         }
 
         let groups: Option<&[u8]>;
-        let regs_read: Option<&[u8]>;
-        let regs_write: Option<&[u8]>;
+        let regs_read: Option<&[u16]>;
+        let regs_write: Option<&[u16]>;
         let extra_detail: Option<&capstone_sys::bindings::cs_detail__bindgen_ty_1>;
         if let Some(detail) = unsafe { insn.detail.as_ref() } {
             groups = Some(&detail.groups[..detail.groups_count as usize]);
